Tree Population and Traversal
This is a program that creates 3 different binary trees. One of them is a tree that is populated with a prefix expression, one of them is a tree that is populated with an infix expression, and one of them is a tree that is populated with a postfix expression. They have a runtime of O(N). Each tree is then printed out in all three orders. Pre-order, in-order, and post-order.